requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy
asyncpg==0.27.0
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
alembic==1.11.1
pytest==7.2.2
pytest-asyncio==0.18.3
httpx==0.24.1